feat: convert recipient lists into strings
For some reason, the framework fails to convert a type instance
that has TStringList-valued property into a json object.
The same seems to be true for TList<String>. Therefore, for the moment
I concatenate the lists of recipients (to, cc and bcc) into a
comma-separated string.
The same is done for list of attachments (it is wrong, since an
attachment is a binary content that might contain a comma).