Adjust tests for the back end responce
When testing how a back end responce object is constructed from a json,
use its string form (like '{"status":true, "msgstat":"ok"}') instead of
using the TJsonObject methods.
This is due to a fact that the json object has a boolean valued key
("status") whose value should be created by means of TBoolTrue.Create
or TBoolfalse.Cteate and not by means of TJsonBool.Create(True) or
TJsonBool.Create(False).
It seems to be a restriction imposed by the DMVCFramework.
In order to get rid of dependency on this issue, let's create a plain
json from a string.