Move UpdatePersistedState from the Add/CancelSubscription
When the method for updating the persisted state is located inside
the methods that manage adding and cancelling the subscriptions, it
means that those methods have some side effects.
When adjusting the tests, I noticed that there were attemps to save the
AQ state in a file (due to the above mentioned side effect). This was not
satisfactory.