Split ActiveQueueAPI in client and consumer parts
ActiveQueueAPI consists in two faces: one for clients and one for
consumers. Since consumers and clients have different dipendencies,
it is better to split ActiveQueueAPi in these parts. Otherwise,
i.e. dispatcher was aware of SubscriptionData class (specific for the
Consumer). It means that there is a dependency that propagates btw
three projects (Dispatcher, AQ and Consumer).